+++ actionpack/test/controller/api/conditional_get_test.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require "abstract_unit"
-require "active_support/core_ext/integer/time"
 require "active_support/core_ext/numeric/time"
 
 class ConditionalGetApiController < ActionController::API