+++ activesupport/lib/active_support/core_ext/integer.rb
@@ -2,4 +2,3 @@
 
 require "active_support/core_ext/integer/multiple"
 require "active_support/core_ext/integer/inflections"
-require "active_support/core_ext/integer/time"