+++ activesupport/lib/active_support/time.rb
@@ -13,7 +13,6 @@ require "active_support/core_ext/time"
 require "active_support/core_ext/date"
 require "active_support/core_ext/date_time"
 
-require "active_support/core_ext/integer/time"
 require "active_support/core_ext/numeric/time"
 
 require "active_support/core_ext/string/conversions"